backend/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20

# Redis / job queue
redis==5.2.1
arq==0.26.1

# AWS S3
boto3==1.35.86

# Claude API
anthropic==0.42.0

# File processing
python-docx==1.1.2
pypdf2==3.0.1

# Email
resend==2.5.0

# Payments
stripe==11.4.1

# Config
pydantic-settings==2.7.1

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1
